add update_view_process
